requirements.txt
python-telegram-bot
aiocryptopay
python-dotenv